+++ plugins/nf-iridanext/src/main/nextflow/iridanext/MetadataParser.groovy
@@ -32,39 +32,4 @@ class MetadataParser {
             return [:]
         }
     }
-
-    public static MetadataParser createMetadataParser(String type, PathMatcher pathMatcher, String id) {
-        if (type == "csv") {
-            return MetadataParserCSV(pathMatcher, id)
-        } else {
-            throw new Exception("Invalid type=${type} for MetadataParser")
-        }
-    }
-}
-
-@CompileStatic
-class MetadataParserCSV extends MetadataParser {
-    public MetadataParserCSV(PathMatcher pathMatcher, String id) {
-        super(pathMatcher, id)
-    }
-
-    @Override
-    public Map<String, Object> parseMetadata(Path path) {
-        return csvToJsonById(path, this.id)
-    }
-
-    private Map<String, Object> csvToJsonById(Path path, String idColumn) {
-        path = Nextflow.file(path) as Path
-        List rowsList = path.splitCsv(header:true, strip:true, sep:',', quote:'\"')
-
-        Map<String, Object> rowsMap = rowsList.collectEntries { row ->
-            if (idColumn !in row) {
-                throw new Exception("Error: column with idColumn=${idColumn} not in CSV ${path}")
-            } else {
-                return [(row[idColumn] as String): (row as Map).findAll { it.key != idColumn }]
-            }
-        }
-
-        return rowsMap
-    }
 }
